Avoid using a 'comma pasting' GNU extension (non-portable, reported my clang).
This is not a 'complete' fix since the DBG macro is still using it; but more globally, this file should not be part of the libnfc installed files.  We should rely on error codes and maybe some logging mechanism (e.g. syslog() on POSIX compliant systems and RegisterEventSource() / ReportEvent() on Windows).



git-svn-id: http://libnfc.googlecode.com/svn/trunk@380 45cb0fa4-f8fe-11dd-984d-ab748d24aa7e